requirements.txt
absl-py==0.7.1
apex==0.1
astor==0.8.0
attrs==19.3.0
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.3.9
cffi==1.12.3
chainer==6.4.0
chainercv==0.13.1
chardet==3.0.4
cloudpickle==1.1.1
colorama==0.4.1
cxxfilt==0.2.0
cycler==0.10.0
Cython==0.29.9
decorator==4.4.0
docker-py==1.10.6
docker-pycreds==0.4.0
docutils==0.14
easydict==1.9
fakeredis==1.0.3
filelock==3.0.12
gast==0.2.2
google==2.0.3
graphviz==0.8.4
grpcio==1.21.1
h5py==2.9.0
horovod==0.16.2
idna==2.8
imageio==2.5.0
importlib-metadata==1.2.0
jmespath==0.9.4
joblib==0.13.2
json-tricks==3.13.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
libkeepass==0.3.0
lxml==4.3.3
Markdown==3.1.1
matplotlib==3.0.3
mock==3.0.5
more-itertools==8.0.1
msgpack==0.6.1
msgpack-numpy==0.4.4.3
munkres==1.1.2
mxnet-cu100mkl==1.4.1
networkx==2.3
numpy==1.16.4
packaging==19.2
pandas==0.24.2
passlib==1.7.1
pathlib2==2.3.5
pg8000==1.13.1
Pillow==6.0.0
pluggy==0.13.1
protobuf==3.10.0
psutil==5.6.2
py==1.8.0
pycocotools==2.0
pycparser==2.19
pycryptodome==3.8.1
pycurl==7.43.0
pydensecrf==1.0rc3
pygobject==3.20.0
pyparsing==2.4.0
pytest==5.3.1
python-apt==1.1.0b1+ubuntu0.16.4.4
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
redis==3.2.1
requests==2.22.0
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.33.11
tabulate==0.8.3
tensorboard==1.13.1
tensorboardX==1.9
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorpack==0.9.0.1
termcolor==1.1.0
thop==0.0.31.post2001170342
torch==1.2.0
torchvision==0.4.1
tqdm==4.32.1
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
yac==1.4.6
yacs==0.1.6
zipp==0.6.0
zmq-ops==0.1.0